squashme
"fix" toctou "bug" that github think is "high severity" because we check
if a file exists before deleting it. There's no point checking, we can
just try to delete it and if it fails, then well the thing didn't exist.

fix log messages with missing new lines

fix missing environment variable notes in --help messages

deconflict CONFDIR/DATADIR/CERTSDIR variable names that are defined in clamav-config.h.in for libclamav from variable that had the same name for use in clamav applications that use the optparser.